refactor(router-core): move getHydrateStatus to router-core for cross-framework reuse
- Relocate hydrate-status from react-router to router-core so Solid/Vue can share it
- Add early break to hydrate-status loop once both explicit true/false flags are set
- Restore conflict warning behind a DEV-only React.useEffect in Scripts
- Document the data-tsr-client-entry marker contract at both ends (emit + strip)
